pyproject.toml
[tool.poetry]
name = "blackneedles"
version = "0.1.0"
description = ""
authors = ["Thiago F Pappacena <pappacena@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
typer = "^0.11.0"
snowflake-connector-python = "^3.7.1"
snowflake-snowpark-python = "^1.14.0"
pydantic = "^2.6.4"
ipython = "^8.22.2"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
blackneedles = "blackneedles.__main__:app"